Implemented feature.
Completed feature where api has open endpoints for listing the collection points stored in the database.
These endpoints use query parameters to filter the collection points needed in the result set.

This commit also has all the basic structure of the project since this was the first development made for this api.